# apps/castaway-web/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Castaway Web API
  version: 0.0.0
...
paths:
...
  /instances:
    get:
      operationId: listInstances
      parameters:
        - name: season
          in: query
          required: false
          schema:
            type: integer
            format: int32
          explode: false
        - name: name
          in: query
          required: false
          schema:
            type: string
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListInstancesResponse'
    post:
      operationId: createInstance
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateInstanceResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInstanceRequest'
  /instances/import:
    post:
      operationId: importInstance
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateInstanceResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportInstanceRequest'
...
  /instances/{instanceID}/contestants:
    post:
      operationId: createContestant
      parameters:
        - name: instanceID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateContestantResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContestantRequest'
    get:
      operationId: listContestants
      parameters:
        - name: instanceID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/ListContestantsResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
...
  /instances/{instanceID}/participants:
    post:
      operationId: createParticipant
      parameters:
        - name: instanceID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateParticipantResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateParticipantRequest'
    get:
      operationId: listParticipants
      parameters:
        - name: instanceID
          in: path
          required: true
          schema:
            type: string
        - name: name
          in: query
          required: false
          schema:
            type: string
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/ListParticipantsResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    Contestant:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
...
    CreateContestantRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
    CreateContestantResponse:
      type: object
      required:
        - contestant
      properties:
        contestant:
          $ref: '#/components/schemas/Contestant'
    CreateInstanceEpisodeRequest:
      type: object
      required:
        - episode_number
        - label
        - airs_at
      properties:
        episode_number:
          type: integer
          format: int32
        label:
          type: string
        airs_at:
          type: string
          format: date-time
    CreateInstanceRequest:
      type: object
      required:
        - name
        - season
      properties:
        name:
          type: string
        season:
          type: integer
          format: int32
        contestants:
          type: array
          items:
            type: string
        managed_progression:
          type: boolean
        episodes:
          type: array
          items:
            $ref: '#/components/schemas/CreateInstanceEpisodeRequest'
    CreateInstanceResponse:
      type: object
      required:
        - instance
      properties:
        instance:
          $ref: '#/components/schemas/Instance'
...
    CreateParticipantRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
    CreateParticipantResponse:
      type: object
      required:
        - participant
      properties:
        participant:
          $ref: '#/components/schemas/Participant'
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...
    ImportInstanceRequest:
      type: object
      required:
        - season
        - submissions
      properties:
        season:
          type: integer
          format: int32
        name:
          type: string
        submissions:
          type: array
          items:
            $ref: '#/components/schemas/ImportSubmission'
    ImportSubmission:
      type: object
      required:
        - participant_name
        - rankings
      properties:
        participant_name:
          type: string
        rankings:
          type: array
          items:
            type: string
    Instance:
      type: object
      required:
        - id
        - name
        - season
        - created_at
      properties:
        id:
          type: string
        name:
          type: string
        season:
          type: integer
          format: int32
        created_at:
          type: string
          format: date-time
        current_episode:
          $ref: '#/components/schemas/InstanceEpisodeBrief'
    InstanceEpisodeBrief:
      type: object
      required:
        - id
        - episode_number
        - label
        - airs_at
      properties:
        id:
          type: string
        episode_number:
          type: integer
          format: int32
        label:
          type: string
        airs_at:
          type: string
          format: date-time
...
    ListContestantsResponse:
      type: object
      required:
        - contestants
      properties:
        contestants:
          type: array
          items:
            $ref: '#/components/schemas/Contestant'
    ListInstancesResponse:
      type: object
      required:
        - instances
      properties:
        instances:
          type: array
          items:
            $ref: '#/components/schemas/Instance'
...
    ListParticipantsResponse:
      type: object
      required:
        - participants
      properties:
        participants:
          type: array
          items:
            $ref: '#/components/schemas/Participant'
...
    Participant:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
        discord_user_id:
          type: string
...
servers:
  - url: http://localhost:8080
    description: Local development
    variables: {}